[tests] refactor fixtures
