+++ tests/test_main.py
@@ -14,18 +14,16 @@ from maestral.utils.path import delete
 import unittest
 from unittest import TestCase
 
-from helpers import (
+from fixtures import (
     setup_test_config,
     cleanup_test_config,
-    acquire_test_lock,
-    release_test_lock,
+    DropboxTestLock
 )
 
 
 @unittest.skipUnless(os.environ.get("DROPBOX_TOKEN"), "Requires auth token")
 class TestAPI(TestCase):
 
-    TEST_LOCK_PATH = "/test.lock"
     TEST_FOLDER_PATH = "/sync_tests"
 
     resources = osp.dirname(__file__) + "/resources"
@@ -33,7 +31,9 @@ class TestAPI(TestCase):
     def setUp(self):
 
         self.m = setup_test_config()
-        acquire_test_lock(self.m, TestAPI.TEST_LOCK_PATH, timeout=60 * 60)
+        self.lock = DropboxTestLock(self.m)
+        if not self.lock.acquire(timeout=60*60):
+            raise TimeoutError("Could not acquire test lock")
 
         # all our tests will be carried out within this folder
         self.test_folder_dbx = TestAPI.TEST_FOLDER_PATH
@@ -51,7 +51,7 @@ class TestAPI(TestCase):
     def tearDown(self):
 
         cleanup_test_config(self.m, self.test_folder_dbx)
-        release_test_lock(self.m, TestAPI.TEST_LOCK_PATH)
+        self.lock.release()
 
     # helper functions
